Organize API source code files' path
