--- src/grpc/protobuf/Makefile
@@ -1,4 +1,0 @@
-# https://github.com/gogo/protobuf/blob/master/test/example/Makefile
-#
-regenerate:
-	(protoc -I=. -I=$(GOPATH)/src/github.com/gogo/protobuf/protobuf -I=$(GOPATH)/src --gogo_out=. cbtumblebug.proto)
